% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_infront_data_files.R
\name{read_infront_data_files}
\alias{read_infront_data_files}
\title{This function read csv Infront data files. Supply both OHLC and info files}
\usage{
read_infront_data_files(data_file, info_file)
}
\arguments{
\item{data_file}{Name and path to OHLC data file}

\item{info_file}{Name and path to file with additional info (e.g gics, description, country, benchmark, currency)}
}
\description{
This function read csv Infront data files. Supply both OHLC and info files
}
